Moved reporting `ros-entries` code into new file
